requirements.txt
streamlit==1.38.0
openai==1.47.0
anthropic==0.34.2
python-dotenv==1.0.1
